# specs/001-task-crud-auth/contracts/tasks.openapi.yaml
openapi: 3.0.3
info:
  title: Tasks API
  description: API for managing user tasks with CRUD operations
  version: 1.0.0
servers:
  - url: http://localhost:8000/api
    description: Local development server
  - url: https://api.example.com
    description: Production server
paths:
  /{user_id}/tasks:
    get:
      summary: Get all tasks for a user
      description: Returns a list of all tasks belonging to the specified user
      tags:
        - Tasks
      parameters:
        - name: user_id
          in: path
          required: true
          description: UUID of the user whose tasks to retrieve
          schema:
            type: string
            format: uuid
        - name: completed
          in: query
          required: false
          description: Filter tasks by completion status
          schema:
            type: boolean
        - name: limit
          in: query
          required: false
          description: Maximum number of tasks to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          required: false
          description: Number of tasks to skip
          schema:
            type: integer
            minimum: 0
            default: 0
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of tasks retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  total:
                    type: integer
                    example: 10
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      summary: Create a new task
      description: Creates a new task for the specified user
      tags:
        - Tasks
      parameters:
        - name: user_id
          in: path
          required: true
          description: UUID of the user creating the task
          schema:
            type: string
            format: uuid
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
            required:
              - title
            properties:
              title:
                type: string
                minLength: 1
                maxLength: 200
                example: "Buy groceries"
              description:
                type: string
                maxLength: 1000
                example: "Milk, eggs, bread, fruits"
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /{user_id}/tasks/{task_id}:
    get:
      summary: Get a specific task
      description: Returns details of a specific task belonging to the user
      tags:
        - Tasks
      parameters:
        - name: user_id
          in: path
          required: true
          description: UUID of the user
          schema:
            type: string
            format: uuid
        - name: task_id
          in: path
          required: true
          description: ID of the task to retrieve
          schema:
            type: integer
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Task retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update a task (full update)
      description: Updates all fields of a specific task belonging to the user
      tags:
        - Tasks
      parameters:
        - name: user_id
          in: path
          required: true
          description: UUID of the user
          schema:
            type: string
            format: uuid
        - name: task_id
          in: path
          required: true
          description: ID of the task to update
          schema:
            type: integer
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update a task (partial update)
      description: Updates specific fields of a task belonging to the user
      tags:
        - Tasks
      parameters:
        - name: user_id
          in: path
          required: true
          description: UUID of the user
          schema:
            type: string
            format: uuid
        - name: task_id
          in: path
          required: true
          description: ID of the task to update
          schema:
            type: integer
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskPatch'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a task
      description: Permanently deletes a specific task belonging to the user
      tags:
        - Tasks
      parameters:
        - name: user_id
          in: path
          required: true
          description: UUID of the user
          schema:
            type: string
            format: uuid
        - name: task_id
          in: path
          required: true
          description: ID of the task to delete
          schema:
            type: integer
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Task deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        title:
          type: string
          maxLength: 200
          example: "Buy groceries"
        description:
          type: string
          maxLength: 1000
          example: "Milk, eggs, bread, fruits"
        completed:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: "2023-01-01T12:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2023-01-01T12:00:00Z"
      required:
        - id
        - user_id
        - title
        - completed
        - created_at
        - updated_at
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: "Buy groceries"
        description:
          type: string
          maxLength: 1000
          example: "Milk, eggs, bread, fruits"
        completed:
          type: boolean
          example: false
      required:
        - title
        - description
        - completed
    TaskPatch:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: "Buy groceries"
        description:
          type: string
          maxLength: 1000
          example: "Milk, eggs, bread, fruits"
        completed:
          type: boolean
          example: true
  responses:
    Unauthorized:
      description: Unauthorized - Invalid or missing JWT token
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Unauthorized"
    NotFound:
      description: Not Found - Resource does not exist or belongs to another user
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Resource not found"
    BadRequest:
      description: Bad Request - Invalid input data
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Invalid input data"